backend/requirements.txt
# AutoAnalytics - Sistema Inteligente para Oficinas
# Dependências principais

# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de Dados & ORM
sqlalchemy==2.0.23

# Autenticação & Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validação de Dados
pydantic==2.5.0
python-multipart==0.0.6

# Manipulação de Arquivos
aiofiles==23.2.1

# Análise de Dados & ML
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
scikit-learn==1.4.0  # ⭐ SUBSTITUI TENSORFLOW!

# Requisições HTTP
requests==2.31.0

# Ambiente & Configuração
python-dotenv==1.0.0
psutil==5.9.6

# Serialização de modelos
joblib==1.3.2

# Para compatibilidade (evitar warnings)
# tensorflow==2.15.0  # COMENTADO - Não funciona sem AVX